Re-organize to move window out of input
Also, re-export Graphics and Input at the top level. They are needed
in literally every Quicksilver application, because you need to name the
types to run 'app.'
