+++ examples/02_image.rs
@@ -2,9 +2,8 @@
 // Draw an image to the screen
 use quicksilver::{
     geom::{Rectangle, Vector},
-    graphics::{Color, Graphics, Image},
-    input::{Input, Window},
-    run, Result, Settings,
+    graphics::{Color, Image},
+    run, Graphics, Input, Result, Settings, Window,
 };
 
 fn main() {
